Complete execution (with incomplete templates)

+++ lib/build.py
@@ -2,21 +2,20 @@
 
 import os
 import sys
+import errno
 import shutil
 import subprocess
 import template
 import urllib
+import zipfile
 
 PATH = os.path.dirname(os.path.realpath(__file__))
 SABHA_PATH = os.path.realpath(os.path.join(PATH, '..', 'bosh-generic-sb-release', 'bin'))
 
-VERBOSE=False
-
 def build(config, verbose=False):
 	context = config.copy()
 	add_defaults(context)
-	global VERBOSE
-	VERBOSE = verbose
+	context['verbose'] = verbose
 	with cd('release', clobber=True):
 		create_bosh_release(context)
 	with cd('product', clobber=True):
@@ -24,39 +23,38 @@ def build(config, verbose=False):
 
 def add_defaults(context):
 	context['stemcell_criteria'] = context.get('stemcell_criteria', {})
-	context['forms'] = context.get('forms', [])
-	context['packages'] = context.get('packages', [])
 
 def create_bosh_release(context):
 	target = os.getcwd()
 	bosh('init', 'release')
 	template.render('src/templates/all_open.json', 'src/templates/all_open.json', context)
 	template.render('config/final.yml', 'config/final.yml', context)
-	bash('fetch_cf_cli.sh', target)
-	pkgs = context.get('packages', [])
-	for pkg in pkgs:
-		jobs = pkg.get('jobs', [])
+	packages = context.get('packages', [])
+	requires_cf_cli = False
+	for package in packages:
+		add_blob_package(context, package)
+		jobs = package.get('jobs', [])
 		cf_push = False
 		cf_create_service_broker = False
 		cf_create_buildpack = False
 		for job in jobs:
 			if job == 'cf-push':
-				cf_push = True
+				requires_cf_cli = True
+				add_bosh_job(context, package, 'deploy-app', post_deploy=True)
+				add_bosh_job(context, package, 'delete-app', pre_delete=True)
 			elif job == 'cf-create-service-broker':
-				cf_create_service_broker = True
+				requires_cf_cli = True
+				add_bosh_job(context, package, 'register-broker', post_deploy=True)
+				add_bosh_job(context, package, 'destroy-broker', pre_delete=True)
 			elif job == 'cf-create-buildpack':
-				cf_create_buildpack = True
+				requires_cf_cli = True
+				add_bosh_job(context, package, 'deploy-buildpack', post_deploy=True)
+				add_bosh_job(context, package, 'delete-buildpack', pre_delete=True)
 			else:
-				print >>sys.stderr, 'unknown job type', job, 'for', pkg.get('name', 'unnamed package')
+				print >>sys.stderr, 'unknown job type', job, 'for', package.get('name', 'unnamed package')
 				sys.exit(1)
-		if cf_push or cf_create_service_broker:
-			bash('addApp.sh', target, pkg['name'], str(cf_push).lower(), str(cf_create_service_broker).lower())
-		if cf_create_buildpack:
-			bash('addBuildpack.sh', target, pkg['name'])
-		bash('addBlob.sh', target, os.path.join('..', pkg['path']), pkg['name'], pkg['name'])
-#	add_cf_cli(context)
-#	add_buildpacks(context)
-#	add_service_brokers(context)
+	if requires_cf_cli:
+		add_cf_cli(context)
 	output = bosh('create', 'release', '--force', '--final', '--with-tarball', '--version', context['version'])
 	context['release'] = bosh_extract(output, [
 		{ 'label': 'name', 'pattern': 'Release name' },
@@ -64,6 +62,87 @@ def create_bosh_release(context):
 		{ 'label': 'manifest', 'pattern': 'Release manifest' },
 		{ 'label': 'tarball', 'pattern': 'Release tarball' },
 	])
+	print
+
+def add_bosh_job(context, package, job_type, post_deploy=False, pre_delete=False):
+	job_name = job_type + '-' + package['name']
+	bosh('generate', 'job', job_name)
+	job_context = {
+		'job_name': job_name,
+		'job_type': job_type,
+		'context': context,
+		'package': package,
+	}
+	template.render(
+		os.path.join('jobs', job_name, 'spec'),
+		os.path.join('jobs', 'spec'),
+		job_context
+	)
+	template.render(
+		os.path.join('jobs', job_name, 'templates', job_name + '.sh.erb'),
+		os.path.join('jobs', job_type + '.sh.erb'),
+		job_context
+	)
+	context['jobs'] = context.get('jobs', []) + [{
+		'name': job_name,
+		'type': job_type,
+		'package': package,
+	}]
+	if post_deploy:
+		context['post_deploy_errands'] = context.get('post_deploy_errands', []) + [{ 'name': job_name }]
+	if pre_delete:
+		context['pre_delete_errands'] = context.get('pre_delete_errands', []) + [{ 'name': job_name }]
+
+def add_src_package(context, package, alternate_template=None):
+	add_package('src', context, package, alternate_template)
+
+def add_blob_package(context, package, alternate_template=None):
+	add_package('blobs', context, package, alternate_template)
+
+def add_package(dir, context, package, alternate_template=None):
+	name = package['name']
+	bosh('generate', 'package', name)
+	target_dir = os.path.realpath(os.path.join(dir, name))
+	package_dir = os.path.realpath(os.path.join('packages', name))
+	mkdir_p(target_dir)
+	template_dir = 'packages'
+	if alternate_template is not None:
+		template_dir = os.path.join(template_dir, alternate_template)
+	package_context = {
+		'context': context,
+		'package': package,
+		'files': []
+	}
+	with cd('..'):
+		for file in package.get('files', []):
+			filename = file.get('name', os.path.basename(file['path']))
+			try:
+				urllib.urlretrieve(file['path'], os.path.join(target_dir, filename))
+			except ValueError: # Invalid URL, assume filename
+				shutil.copy(os.path.join('..', file['path']), target_dir)
+			package_context['files'] += [ filename ]
+	template.render(
+		os.path.join(package_dir, 'spec'),
+		os.path.join(template_dir, 'spec'),
+		package_context
+	)
+	template.render(
+		os.path.join(package_dir, 'packaging'),
+		os.path.join(template_dir, 'packaging'),
+		package_context
+	)
+
+def add_cf_cli(context):
+	add_blob_package(context,
+		{
+			'name': 'cf_cli',
+			'files': [{
+				'name': 'cf-linux-amd64.tgz',
+				'path': 'https://cli.run.pivotal.io/stable?release=linux64-binary&source=github-rel'
+			}]
+		},
+		alternate_template='cf_cli'
+	)
 
 def bash(*argv):
 	argv = list(argv)
@@ -81,23 +160,21 @@ def bash(*argv):
 def create_tile(context):
 	release = context['release']
 	release['file'] = os.path.basename(release['tarball'])
-#	with cd('releases'):
-#		print 'tile generate release'
-#		shutil.copy(release['tarball'], release['file'])
+	with cd('releases'):
+		print 'tile generate release'
+		shutil.copy(release['tarball'], release['file'])
 	print 'tile generate metadata'
 	template.render('metadata/' + release['name'] + '.yml', 'tile/metadata.yml', context)
-#	with cd('content_migrations'):
-#		print 'tile generate content-migrations'
-#		migrations = tile_migrations(context, release)
-#		with open(release['name'] + '.yml', 'wb') as f:
-#			write_yaml(f, migrations)
-#	pivotal_file = release['name'] + '-' + release['version'] + '.pivotal'
-#	with zipfile.ZipFile(pivotal_file, 'w') as f:
-#		f.write(os.path.join('releases', release['file']))
-#		f.write(os.path.join('metadata', release['name'] + '.yml'))
-#		f.write(os.path.join('content_migrations', release['name'] + '.yml'))
-#	print
-#	print 'created tile', pivotal_file
+	print 'tile generate content-migrations'
+	template.render('content_migrations/' + release['name'] + '.yml', 'tile/content-migrations.yml', context)
+	print 'tile package'
+	pivotal_file = release['name'] + '-' + release['version'] + '.pivotal'
+	with zipfile.ZipFile(pivotal_file, 'w') as f:
+		f.write(os.path.join('releases', release['file']))
+		f.write(os.path.join('metadata', release['name'] + '.yml'))
+		f.write(os.path.join('content_migrations', release['name'] + '.yml'))
+	print
+	print 'created tile', pivotal_file
 
 def bosh_extract(output, properties):
 	result = {}
@@ -169,7 +246,7 @@ class cd:
 
     def __enter__(self):
         self.savedPath = os.getcwd()
-        if os.path.isdir(self.newPath):
+        if self.clobber and os.path.isdir(self.newPath):
 			shutil.rmtree(self.newPath)
         mkdir_p(self.newPath)
         os.chdir(self.newPath)
